Minimum viable product
Two classes, one which implements an UDP listener, and the other a UDP
sender.  The sender broadcasts a UDP package comprised of an ID and an
IP End Point to a predefined port. (The IP End Point signifies an end
point on the sender which the receiving application is intended to open
future communication to).
The listener spawns an additional thread and listens for UDP packages on
a predefined port.  When a package is received it is parsed and
validated.  On successful validation, the listener raises an event with
the IP End Point.

A sample console application shows how the listener can be started and
the sender activated.